move set_sqlite_llm_cache
