+++ bot/llm.py
@@ -1,7 +1,11 @@
 import functools
 import os
+from pathlib import Path
 
+from langchain.globals import set_llm_cache
+from langchain_community.cache import SQLiteCache
 from langchain_core.language_models.chat_models import BaseChatModel
+from loguru import logger
 
 
 @functools.cache
@@ -16,3 +20,12 @@ def get_llm_from_env() -> BaseChatModel:
         return ChatGoogleGenerativeAI(model="gemini-1.5-flash", temperature=0)
     else:
         raise ValueError("No API key found in environment variables")
+
+
+def set_sqlite_llm_cache() -> None:
+    database_path = Path.home() / ".cache" / ".langchain.db"
+    logger.info("Using SQLite cache: {}", database_path)
+
+    cache = SQLiteCache(database_path.as_posix())
+
+    set_llm_cache(cache)
